Move all messages to one file as cosmos convention
